Support Rotation Animation on Spinner (mac)
